Done for second review
divided the single file in to multiple according to the classes.
