+++ ConnectFourCPPGame/DefineValues.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#define ROW 6
+#define COLUMN 7
+#define TOTAL_CONNECT 4
+#define PLACEHOLDER_CHARACTER 'O'
