+++ ConnectFourCPPGame/Game.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "MainGameLoop.h"
+
+class Game {
+private:
+    MainGameLoop mainGameLoop;
+public:
+    Game();
+    void StartGame();
+
+};
